samples/GCP/ProcessEngineering/requirements.txt
# ADK runtime
google-adk>=0.1.42
google-auth>=2.17.3

# Web App
flask>=3.0.0

# Data Handling
pandas
numpy
json-repair>=0.30.0

# Document Generation
python-docx>=1.1.0

# Pure-Python Diagram Generation
networkx>=3.0
matplotlib>=3.8.0
scipy

# Security updates
starlette>=0.49.1
